# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "abacus",
    "description": "",
    "license": {
      "name": "EUPL-1.2",
      "identifier": "EUPL-1.2"
    },
    "version": "1.1.0"
  },
  "paths": {
    "/api/account": {
      "get": {
        "summary": "Get current logged-in user endpoint (administrator, coordinator_gsb, coordinator_csb, typist_gsb, typist_csb)",
        "operationId": "account",
        "responses": {
          "200": {
            "description": "The current user name and id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid user session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "administrator",
              "coordinator_gsb",
              "coordinator_csb",
              "typist_gsb",
              "typist_csb"
            ]
          }
        ]
      },
      "put": {
        "summary": "Update the user's account with a new password and optionally new fullname (administrator, coordinator_gsb, coordinator_csb, typist_gsb, typist_csb)",
        "operationId": "account_update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The logged in user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": [
              "administrator",
              "coordinator_gsb",
              "coordinator_csb",
              "typist_gsb",
              "typist_csb"
            ]
          }
        ]
      }
    },
...
    "/api/initialise/admin-exists": {
      "get": {
        "tags": [
          "public"
        ],
        "operationId": "admin_exists",
        "responses": {
          "204": {
            "description": "First admin user exists"
          },
          "403": {
            "description": "Forbidden, the application is already initialised",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No admin user exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/initialise/first-admin": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Create the first admin user, only allowed if no users exist yet",
        "operationId": "create_first_admin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFirstAdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "First admin user created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden, an active user already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (username already exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/initialised": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Check whether the application is initialised (an admin user exists + has logged in at least once)",
        "operationId": "initialised",
        "responses": {
          "204": {
            "description": "The application is initialised"
          },
          "418": {
            "description": "The application is not initialised",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/login": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Login endpoint, authenticates a user and creates a new session + session cookie",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The logged in user id and user name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Logout endpoint, deletes the session cookie",
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Successful logout, or user was already logged out"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountUpdateRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "fullname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CreateFirstAdminRequest": {
        "type": "object",
        "required": [
          "username",
          "fullname",
          "temp_password"
        ],
        "properties": {
          "fullname": {
            "type": "string"
          },
          "temp_password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Credentials": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorReference": {
        "type": "string",
        "description": "Error reference used to show the corresponding error message to the end-user",
        "enum": [
          "AirgapViolation",
          "AlreadyInitialised",
          "ApportionmentAllListsExhausted",
          "ApportionmentCommitteeSessionNotCompleted",
          "ApportionmentDrawingOfLotsRequired",
          "ApportionmentZeroVotesCast",
          "CommitteeSessionPaused",
          "DatabaseError",
          "DataEntryAlreadyClaimed",
          "DataEntryAlreadyFinalised",
          "DataEntryCannotBeDeleted",
          "DataEntryGetNotAllowed",
          "DataEntryNotAllowed",
          "EmlImportError",
          "EmlError",
          "EntryNotFound",
          "EntryNotUnique",
          "Forbidden",
          "InternalServerError",
          "InvalidCommitteeSessionStatus",
          "InvalidData",
          "InvalidHash",
          "InvalidJson",
          "InvalidPassword",
          "InvalidPoliticalGroup",
          "InvalidSession",
          "InvalidStateTransition",
          "InvalidUsernameOrPassword",
          "InvalidVoteCandidate",
          "InvalidVoteGroup",
          "InvalidXml",
          "InvestigationHasDataEntryOrResult",
          "InvestigationRequiresCorrectedResults",
          "NotInitialised",
          "OwnAccountCannotBeDeleted",
          "PasswordRejectionSameAsOld",
          "PasswordRejectionSameAsUsername",
          "PasswordRejectionTooShort",
          "PdfGenerationError",
          "PollingStationRepeated",
          "PollingStationValidationErrors",
          "RequestPayloadTooLarge",
          "Unauthorized",
          "UsernameNotUnique",
          "UserNotFound"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Response structure for errors",
        "required": [
          "error",
          "fatal",
          "reference"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "fatal": {
            "type": "boolean"
          },
          "reference": {
            "$ref": "#/components/schemas/ErrorReference"
          }
        },
        "additionalProperties": false
      },
...
      "LoginResponse": {
        "type": "object",
        "required": [
          "user_id",
          "username",
          "role",
          "needs_password_change"
        ],
        "properties": {
          "fullname": {
            "type": "string"
          },
          "needs_password_change": {
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "user_id": {
            "$ref": "#/components/schemas/UserId"
          },
          "username": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "administrator",
          "coordinator_gsb",
          "coordinator_csb",
          "typist_gsb",
          "typist_csb"
        ]
      },
...
      "User": {
        "type": "object",
        "description": "User object, corresponds to a row in the users table",
        "required": [
          "id",
          "username",
          "role",
          "updated_at",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "fullname": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/UserId"
          },
          "last_activity_at": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "updated_at": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserId": {
        "type": "integer",
        "format": "int32",
        "minimum": 0
      },
...
    },
    "securitySchemes": {
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ABACUS_SESSION"
      }
    }
  }
}